[all] Removing portal and sticky components.  Using general popover
