+++ packages/@sanity/components/src/dialogs/PopOver.js
@@ -1,152 +1,129 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import styles from './styles/PopOver.css'
+import styles from 'part:@sanity/components/dialogs/popover-style'
+import Button from 'part:@sanity/components/buttons/default'
 import CloseIcon from 'part:@sanity/base/close-icon'
-import StickyPortal from 'part:@sanity/components/portal/sticky'
+import {Portal} from 'react-portal'
 import Stacked from '../utilities/Stacked'
 import CaptureOutsideClicks from '../utilities/CaptureOutsideClicks'
 import Escapable from '../utilities/Escapable'
+import {Manager, Target, Popper, Arrow} from 'react-popper'
 
-const PADDING = 20
-
-export default class PopOver extends React.Component {
+export default class EditItemPopOver extends React.Component {
 
   static propTypes = {
+    title: PropTypes.string,
     children: PropTypes.node.isRequired,
     onClose: PropTypes.func,
-    isOpen: PropTypes.bool,
-    color: PropTypes.oneOf(['default', 'danger', 'success', 'warning', 'info']),
-    useOverlay: PropTypes.bool
+    onAction: PropTypes.func,
+    actions: PropTypes.arrayOf(PropTypes.shape({
+      kind: PropTypes.string,
+      title: PropTypes.string,
+      key: PropTypes.string
+    }))
   }
 
   static defaultProps = {
-    color: 'default',
-    onClose() {}, // eslint-disable-line
-    isOpen: true,
-    useOverlay: true
+    title: undefined,
+    onClose() {},
+    onAction() {},
+    actions: [],
+    isOpen: true
   }
 
-  state = {
-    arrowLeft: 0,
-    popoverLeft: 0
+  setTargetRef = element => {
+    this.target = element
+    return element
   }
 
-  handleClose = () => {
-    this.props.onClose()
-  }
-
-  setArrowElement = element => {
-    this._arrowElement = element
-  }
-
-  setContentElement = element => {
-    this._contentElement = element
-  }
-
-  setPopoverInnerElement = element => {
-    this._popOverInnerElement = element
-  }
-
-  setRootElement = element => {
-    this._rootElement = element
-  }
-
-  handlePortalResize = dimensions => {
-    if (!this._popOverInnerElement) {
-      return
-    }
-
-    const {
-      rootLeft,
-      availableHeight,
-      availableWidth
-    } = dimensions
-
-    const width = this._popOverInnerElement.offsetWidth
-
-    let popoverLeft = rootLeft - (width / 2)
-
-    if (availableWidth < (rootLeft + (width / 2))) {
-      popoverLeft = availableWidth - width - PADDING
-    }
-
-    this.setState({
-      popoverLeft: popoverLeft,
-      availableHeight: availableHeight,
-      arrowLeft: rootLeft
-    })
+  setPopperRef = element => {
+    this.popper = element
+    return element
   }
 
   render() {
     const {
+      title,
       children,
-      isOpen,
+      actions,
       onClose,
-      color,
-      useOverlay
     } = this.props
 
-
-    const {
-      popoverLeft,
-      arrowLeft,
-      availableHeight,
-    } = this.state
-
     return (
-      <div style={{display: 'span'}} ref={this.setRootElement}>
-        <StickyPortal
-          isOpen={isOpen}
-          onResize={this.handlePortalResize}
-          useOverlay={useOverlay}
-        >
-          <Stacked>
-            {isActive => (
-              <div
-                ref={this.setPopoverInnerElement}
-                className={`
-                  ${styles.root}
-                  ${color === 'danger' ? styles.colorDanger : ''}
-                  ${color === 'warning' ? styles.colorWarning : ''}
-                  ${color === 'info' ? styles.colorInfo : ''}
-                  ${color === 'success' ? styles.colorSuccess : ''}
-                `}
-              >
-                <div
-                  className={styles.arrow}
-                  ref={this.setArrowElement}
-                  style={{
-                    left: `${arrowLeft}px`
-                  }}
-                />
-                <Escapable onEscape={event => ((isActive || event.shiftKey) && onClose())} />
-                <CaptureOutsideClicks onClickOutside={isActive ? onClose : null}>
-                  <div
-                    className={styles.popover}
-                    style={{
-                      left: `${popoverLeft}px`
-                    }}
-                  >
-                    <button className={styles.close} type="button" onClick={onClose}>
-                      <CloseIcon />
-                    </button>
-
-                    <div
-                      ref={this.setContentElement}
-                      className={styles.content}
-                      style={{
-                        maxHeight: `${availableHeight - PADDING}px`
-                      }}
-                    >
-                      {children}
+      <Stacked>
+        {isActive => (
+          <div>
+            <Escapable onEscape={event => ((isActive || event.shiftKey) && onClose())} />
+            <Manager>
+              <Target className={styles.target} innerRef={this.setTargetRef} />
+              <Portal>
+                <div className={styles.overlay} />
+                <Popper
+                  innerRef={this.setPopperRef}
+                  className={styles.popper}
+                  placement="auto"
+                  modifiers={
+                    {
+                      flip: {
+                        boundariesElement: 'viewport'
+                      },
+                      preventOverflow: {
+                        boundariesElement: 'viewport'
+                      }
+                    }
+                  }
+                >
+                  <Arrow className={title ? styles.filledArrow : styles.arrow} />
+                  <CaptureOutsideClicks onClickOutside={isActive ? onClose : null}>
+                    <div className={styles.popover}>
+                      <button className={title ? styles.closeInverted : styles.close} type="button" onClick={onClose}>
+                        <CloseIcon />
+                      </button>
+                      {
+                        title && (
+                          <h3 className={styles.title}>
+                            {title}
+                          </h3>
+                        )
+                      }
+                      <div className={styles.content}>
+                        {children}
+                      </div>
+                      {
+                        actions.length > 0 && (
+                          <div className={styles.footer}>
+                            <div className={styles.actions}>
+                              {
+                                actions.map((action, i) => {
+                                  return (
+                                    <Button
+                                      key={i}
+                                      onClick={() => this.props.onAction(action)}
+                                      data-action-index={i}
+                                      color={action.color}
+                                      disabled={action.disabled}
+                                      kind={action.kind}
+                                      autoFocus={action.autoFocus}
+                                      className={action.secondary ? styles.actionSecondary : ''}
+                                    >
+                                      {action.title}
+                                    </Button>
+                                  )
+                                })
+                              }
+                            </div>
+                          </div>
+                        )
+                      }
                     </div>
-                  </div>
-                </CaptureOutsideClicks>
-              </div>
-            )}
-          </Stacked>
-        </StickyPortal>
-      </div>
+                  </CaptureOutsideClicks>
+
+                </Popper>
+              </Portal>
+            </Manager>
+          </div>
+        )}
+      </Stacked>
     )
   }
 }
