+++ packages/@sanity/components/src/edititem/story.js
@@ -3,43 +3,13 @@ import React from 'react'
 import {storiesOf, action} from 'part:@sanity/storybook'
 import EditItemPopOver from 'part:@sanity/components/edititem/popover'
 import EditItemFold from 'part:@sanity/components/edititem/fold'
-import Button from 'part:@sanity/components/buttons/default'
 import Sanity from 'part:@sanity/storybook/addons/sanity'
 import Chance from 'chance'
 const chance = new Chance()
 import {withKnobs, boolean, text, number} from 'part:@sanity/storybook/addons/knobs'
 
-const overflowHidden = {
-  minHeight: '100vh',
-  maxWidth: '50vw',
-  margin: '0 auto',
-  position: 'relative',
-  overflow: 'hidden',
-  boxSizing: 'border-box',
-  border: '1px solid red'
-}
-
 storiesOf('Edit item')
   .addDecorator(withKnobs)
-  .add(
-    'PopOver',
-    () => {
-      return (
-        <div style={overflowHidden} id="myScrollContainerId">
-        Things is in the background here.
-          <Button onClick={action('oh noes! I should not ble clickable!')}>Try click me</Button>
-          <Sanity part="part:@sanity/components/edititem/popover" propTables={[EditItemPopOver]}>
-            <EditItemPopOver
-              title={text('title (prop)', 'Edit this item')}
-              onClose={action('onClose')}
-            >
-              Put your form here
-            </EditItemPopOver>
-          </Sanity>
-        </div>
-      )
-    }
-  )
   .add(
     'Fold',
     () => {
