--- packages/@sanity/components/src/portal/story.js
@@ -1,220 +1,0 @@
-/* eslint-disable max-len */
-import React from 'react'
-import StickyPortal from 'part:@sanity/components/portal/sticky'
-import {storiesOf, action} from 'part:@sanity/storybook'
-import {withKnobs, boolean} from 'part:@sanity/storybook/addons/knobs'
-import Sanity from 'part:@sanity/storybook/addons/sanity'
-
-const SMALLTEXT = 'Da vup vefwepi roz ocaljog enkubit buhnuze pu ahubin nu fabvioce sowzopop wa.'
-
-const TEXT = (
-  <div>
-    Sebo emseun fotevza gop gitdurjin gopojna zov azi mu bumfa liv ta bezelocit cuetivu. Bag tonpa si waki ducira ga jizuwet nebnal jo rorug cevcade hiko gemsa liw neco numseg upwug. Tef ivi deneca zufbu us sosfu vub ofetosgum ucimetol isika copsaglof weojkiv fondiju.
-    Da vup vefwepi roz ocaljog enkubit buhnuze pu ahubin nu fabvioce sowzopop wa. Tehwuku ugaon cusnonosi kiocesep osaesego cu hasala aboecto pomoz zanog vifniuse higwawaw enovebar rumeju. Ofta nukpaw keaz focedata va du hira ujsaut cucoc powahorot ra bisruwpa likummec lejnorhi mozemu acofu ma. Pipzu oroje vat es betdiv evetem mehrijjet nuhsebo girocmu ejain gowen pos. Cethap hegpi tibmeheju mana pekguim karaz edebu kone da juc uha ibunakomi foka ke hucalu gu wezeb. Pupmo epetivfe wof be ozopuraw ite oziju lo kac haave koes pivu wubeb jevnovuk bodewfos.Tisehlag muwu asbenef fij haruz ris duhrav wo jevatolo ege osepago tit toh wugfohna jit nisevfo pehketfoh copnusin. Wesdu wuntiduv nujzok ajili leeb pipag wiicunos kouraan uctofzi dihlap pam ro. Uthozlol daamiko aver ka menawomi udtebku zu epa mejziv soihu ganuz del tutim vumwu. Uwoisezi lisja vag upa rohgenas gufced kugtariz av tucog ricracem gojvuh ekralrap cuhosje umlum. Cemeg va ahi obahi fa heanuaz toj he voduz fim omazupsag efo tu is ke.Pu racauke locuzsu jic ugwaci bopcovapa is ja okojepkec wofbat od hi not new acmusab bi. Ki zi low rek rifi deovudiw tip asiucesig tujwotte cafrobir betpulec togocom tab veziawu puvzivhad vegugvo luajcod. Nafaaga takbew goha fer et lo dorerom dopom ju timsoecu tov ozlazeh. Ugi awodva usibo hukrojbe ak fouke ceppahkaj sojkulut sec wuku epeogvaz hajbem lined falubfej zozkobma biveso ruzdo bubazoap. Tefnadod goafosim zihawtip tifnewe uh hajukfu vamotdi we adizu ut evoowutu we wioji coim.Zelpowu gejunsuv vugu garciro woake bab liud puuke do nuesan jak joewboc ujicohhu womigpit ejiveb beb ku. Cetu hisirum tugav akuvew wa vidriffov lowmid zogruwib mod lerba je utsu nilnuboh irajinuc. Ijgu uwu repewu soz meta du laduz pozofgoz mefudu zor kicrulla fiwowwid man ja itoasuke zucpifu buctuzteg nirpi. Apokeek bun ma legecep bulcafec pu bamlam lelibi luasdot biwud feunapi umtir jez ninaggo ahji jul amjehic. Gomi jo ded ibpah diize bev ev af uwkejit guphe siblun oz joligko naedaiv buocjon. Remacod cagolwo vot sit gowhaf bibga or ge te je vuucutes ir be.One zuw bejefwi bu uc opduk nema ri zuh meho porsaj hobbeija wo edioremag runam botiz lavim. Cisuwhak wun wicvupe hedcin fojfu ar onolo lu wobgi lo dus vileaw cugro nuk. Ra wis pi po sejuh hegti uc cozcuwsic gol ika bakaj sof ubo gu. Feb odo wa woci doebaga fumoztuz po buveazi kiezohol iveus ciuj on gawe emanicgu vol mieba. Hano foskoh rivo keb tekuar zamran cafumag gigonsa kowevo websazum rik suvzopori. Hagetso dituba nawafepu jov ek awajawwe cuilo wumkalet fubnu esa pirkacfo azje toutas remul kiicaja naw rokjat. Bijuwa ijo josu napivrab ocelitet ubuodu logweje da ra cebzabuw gorru po jeijke mu.Gutkop ubeit duke zeakoeru dawfab vevec ulo conaep ak umoknig ohueza nez gipi pa. Inedeut wergozaga wo cep upvewo sabzadmo mahsewen rarwidad vo odin fiznurin taemega tufa hif. Rov amnod gecnag asikijiwi ej ze dokil mef cak eccojhib tubemena aguonomel.Codok piomodag utumufac zu bibne azihucbec woje awwu lokalhiw soreza cehbe ic je ubeofom. Pehu atetac uva suzwa tuwkosa nem guk oso ve vahej teotsow il. Ro bazzealu ri legjume gahcuk debi ip reg ocgikmez gudreb gen ulo icuon dethorfaz sogmuzgim geup zi. Guepuos kav usu lem sijve rotiwrag faip rekogmuj binkeni vit difiluk sufhali nanicu ogioh omidac dawe zalo.Jinituw arade kameuk wakuene nidge lejoripo ruma geege notpi rugi desog vependa ehef suv rioz tala ihbumi. Ceztuf pewfoeh zuv ti bos dojiw pelviwje davcahud tir imasuh op jok wihuk. Odoj rustamsaj de totlun abasu sifi lurkodbuz mop dic duc ozaojace eza taw so fu roflin ipbup zodbojo. Wek ro amepiepo pim icjumu ambobo dif deib da mansi ab tahu kuwer dovevi ajsud kogde. Tikjivwi ohiwog get toni wupaw inuluwi lirkumi vab kulta tilfokje jiwbebidu zaneplo jo vuc cojereca ubi to. Acezegda er ugogecec otguknu ujoujawik ivacjar rip ilihaez rifdambo eddu nehbelol his feb. Saun wec bupvif votcujico watmuhhu uponefofa wi deru wabulid peh onu ihebe gutottuh kuloh.Asa ofcosek few dupiga rinlewtu vatnobkul he ak itse or gu naju emmi jafimur vuzwoj jogok. Age nudzagoh hahikiz he jor lunhaspo bizba hofseva jek pahcum zuawu oje dosvuros viwogego jevozbo wovowuh. Vo tod gishicu huuli had sotik set sogojkod mojatuju iwewujov hez eg hib. Teru ankiga ugisa ma be wapror hasetroc rapawuvof dubkig bubeki raba lez sib kot tub. Zen fejti cefbo ato tueboek zozarzaj utaj cuguv gobcu et fo de daje. Kub camepap ra gu bowlap ita cuvjej apamera joj zilwamob kuj pamutu. Fil cuhol udsiw led ahifoec rujibi godzep uh mecsezil ifipe waksi er hirsah bafmajti.Rowenojo mazuddoh ufuwareg iwuivehu saget era godbebi celel jako delzi vudvotboh taduz oletivof rokjoumi. Wap owoobe rop duipjew cadtivu deh pikaiji mijis lijlam cupo kubojita eb jutsu okgufu uhu. Riasaved lute be edubihew upa ovmijif vobpigcat zejoj cohigud romtalsu epcuve kaf. Dil javwine je nisucu ezu izejsu meef cok uwekarvu hoween ki beg recu ho toksumute poz lilreko iwifinu.Moucugel ovu ohwe ke kehgotru kavisi di duhetnog heibu cifkuemi cuz ma mig. Dag huobuge migafwi zuvfaz mihem coviwpe sorevan gumpupi uzubju ahpon hulurnim bawtirmen ha eruecaf. Izeuriaha bal nupum dehaaw uf tahgo lecefjoc owadi new ziotre givfiseb odi bam oraewbo zubi guw pocpotuvi mina.
-  </div>
-)
-
-const showResize = function (resizeObject) {
-  //console.log(resizeObject)
-}
-
-storiesOf('Portal')
-  .addDecorator(withKnobs)
-  .add(
-    'Sticky',
-    () => {
-
-      const scrollStyle = {
-        width: '70vw',
-        height: '70vh',
-        border: '1px dotted #ccc',
-        position: 'relative',
-        overflow: 'scroll'
-      }
-
-      return (
-        <Sanity part="part:@sanity/components/portal/sticky" propTables={[StickyPortal]}>
-          <div
-            style={scrollStyle}
-            id="normalStickyScrollId"
-          >
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            <span style={{border: '1px solid red'}}>ROOT
-              <StickyPortal
-                onlyBottomSpace={boolean('onlyBottomSpace (prop)', true)}
-                isOpen={boolean('isOpen (prop)', false)}
-                onClose={action('onClose')}
-                stickToTop={boolean('stickToTop (prop)', false)}
-                scrollContainer={document.getElementById('normalStickyScrollId')}
-                onResize={showResize}
-              >
-                <div
-                  style={{
-                    display: 'block',
-                    border: '1px solid magenta'
-                  }}
-                >
-                Content
-                </div>
-              </StickyPortal>
-            </span>
-            {TEXT}
-          </div>
-        </Sanity>
-      )
-    }
-  )
-
-  .add(
-    'Sticky (nested)',
-    () => {
-
-      const scrollStyle = {
-        width: '70vw',
-        height: '70vh',
-        border: '1px dotted #ccc',
-        position: 'relative',
-        overflow: 'scroll'
-      }
-
-      return (
-        <Sanity part="part:@sanity/components/portal/sticky" propTables={[StickyPortal]}>
-          <div
-            style={scrollStyle}
-          >
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            {SMALLTEXT}
-            <span style={{border: '1px solid red'}}>ROOT
-              <StickyPortal
-                key="stickyNested1"
-                isOpen={boolean('1: isOpen (prop)', false)}
-                stickToTop={boolean('1: stickToTop (prop)', false)}
-                onClose={action('onClose')}
-              >
-                <div
-                  style={{
-                    overflow: 'scroll',
-                    height: '200px',
-                    width: '200px',
-                    border: '1px solid red'
-                  }}
-                >
-                  {TEXT}
-                ROOT -
-                  <StickyPortal
-                    key="stickyNested2"
-                    isOpen={boolean('2: isOpen (prop)', false)}
-                    stickToTop={boolean('2: stickToTop (prop)', false)}
-                    onClose={action('onClose')}
-                    onResize={showResize}
-                  >
-                    <div
-                      style={{
-                        border: '10px solid red',
-                        backgroundColor: 'red',
-                        color: 'white'
-                      }}
-                    >
-                      {SMALLTEXT}
-                      {SMALLTEXT}
-                    </div>
-                  </StickyPortal>
-                  {SMALLTEXT}
-                </div>
-              </StickyPortal>
-            </span>
-            {TEXT}
-          </div>
-        </Sanity>
-      )
-    }
-  )
-
-  .add(
-    'Sticky (simple)',
-    () => {
-
-      const scrollStyle = {
-        width: '80vw',
-        height: '80vh',
-        border: '1px dotted #ccc',
-        position: 'relative',
-        overflow: 'scroll'
-      }
-
-      return (
-        <Sanity part="part:@sanity/components/portal/sticky" propTables={[StickyPortal]}>
-          <div
-            style={scrollStyle}
-          >
-            {SMALLTEXT}
-            <StickyPortal
-              isOpen={boolean('isOpen (prop)', false)}
-              onClose={action('onClose')}
-              stickToTop={boolean('stickToTop (prop)', false)}
-            >
-              <div
-                style={{
-                  border: '1px solid magenta',
-                  backgroundColor: 'rgba(120,0,120,0.3)',
-                  position: 'abolute',
-                  top: '0',
-                  left: '0',
-                  bottom: '0',
-                  width: '300px',
-                  height: '100%',
-                  overflow: 'scroll'
-                }}
-              >
-                {TEXT}
-              </div>
-            </StickyPortal>
-            {TEXT}
-          </div>
-        </Sanity>
-      )
-    }
-  )
-
-  .add(
-    'Sticky without scrollctonainer',
-    () => {
-
-      return (
-        <Sanity part="part:@sanity/components/portal/sticky" propTables={[StickyPortal]}>
-          <div>
-            {SMALLTEXT}
-            <StickyPortal
-              isOpen={boolean('isOpen (prop)', false)}
-              onClose={action('onClose')}
-              stickToTop={boolean('stickToTop (prop)', false)}
-            >
-              <div
-                style={{
-                  border: '1px solid magenta',
-                  backgroundColor: 'rgba(120,0,120,0.3)',
-                  position: 'abolute',
-                  top: '0',
-                  left: '0',
-                  bottom: '0',
-                  width: '300px',
-                  height: '50%',
-                  overflow: 'scroll'
-                }}
-              >
-                {TEXT}
-              </div>
-            </StickyPortal>
-            {TEXT}
-          </div>
-        </Sanity>
-      )
-    }
-  )
